+++ limen/indicators/__init__.py
@@ -1,6 +1,9 @@
 from limen.indicators.atr import atr
+from limen.indicators.avgprice import avgprice
 from limen.indicators.body_pct import body_pct
 from limen.indicators.macd import macd
+from limen.indicators.medprice import medprice
+from limen.indicators.midprice import midprice
 from limen.indicators.ppo import ppo
 from limen.indicators.price_change_pct import price_change_pct
 from limen.indicators.returns import returns
@@ -8,10 +11,17 @@ from limen.indicators.roc import roc
 from limen.indicators.rolling_volatility import rolling_volatility
 from limen.indicators.rsi_sma import rsi_sma
 from limen.indicators.sma import sma
+from limen.indicators.typprice import typprice
+from limen.indicators.var import var
 from limen.indicators.wilder_rsi import wilder_rsi
+from limen.indicators.wclprice import wclprice
 from limen.indicators.bollinger_bands import bollinger_bands
 from limen.indicators.bollinger_position import bollinger_position
 from limen.indicators.cci import cci
+from limen.indicators.linearreg import linearreg
+from limen.indicators.linearreg_angle import linearreg_angle
+from limen.indicators.linearreg_intercept import linearreg_intercept
+from limen.indicators.linearreg_slope import linearreg_slope
 from limen.indicators.stochastic_oscillator import stochastic_oscillator
 from limen.indicators.sma_deviation_std import sma_deviation_std
 from limen.indicators.midpoint import midpoint
@@ -19,12 +29,19 @@ from limen.indicators.window_return import window_return
 
 __all__ = [
     'atr',
+    'avgprice',
     'body_pct',
     'bollinger_bands',
     'bollinger_position',
     'cci',
+    'linearreg',
+    'linearreg_angle',
+    'linearreg_intercept',
+    'linearreg_slope',
     'macd',
+    'medprice',
     'midpoint',
+    'midprice',
     'ppo',
     'price_change_pct',
     'returns',
@@ -34,6 +51,9 @@ __all__ = [
     'sma',
     'sma_deviation_std',
     'stochastic_oscillator',
+    'typprice',
+    'var',
+    'wclprice',
     'wilder_rsi',
-    'window_return'
+    'window_return',
 ]